requirements.txt
appdirs==1.4.3
cached-property==1.4.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.9
defusedxml==0.5.0
idna==2.6
isodate==0.6.0
python-cli-ui==0.6.1
python-dotenv==0.8.0
pytz==2018.3
requests==2.18.4
requests-toolbelt==0.8.0
six==1.11.0
tabulate==0.8.2
Unidecode==1.0.22
urllib3==1.22
xmltodict==0.11.0
